# Conflicts:
#	Animation_Utilities.h
#	Array_String.cpp
#	Array_String.h
#	Assets/Shaders/duration_spiral.frag
#	General_Purpose.cpp
#	General_Purpose.h
#	Globals.cpp
#	Render.cpp
#	Render.h
#	SDL_Imgui_State.cpp
#	Scene_Graph.cpp
#	Scene_Graph.h
#	Session.cpp
#	Spell.cpp
#	Spell.h
#	Warg_Common.cpp
#	Warg_Common.h
#	Warg_Server.h
#	Warg_State.cpp
#	Warg_State.h
#	main.cpp

+++ Serialization_Buffer.h
@@ -192,5 +192,5 @@ private:
     if (write_ptr)
       memcpy(write_ptr, &data[peek_index], size);
   }
-  uint32 read_index = 0;
+  size_t read_index = 0;
 };
